# pkg/client/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: info@green-ecolution.de
    name: Green Ecolution
    url: https://green-ecolution.de
  description: This is the API for the Green Ecolution Management System.
  license:
    name: AGPL
    url: https://raw.githubusercontent.com/green-ecolution/green-ecolution-management/develop/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Green Space Management API
  version: develop
servers:
- url: /
paths:
...
  /v1/watering-plan:
    get:
      description: Get all watering plans
      operationId: get-all-watering-plans
      parameters:
      - description: Page
        in: query
        name: page
        schema:
          type: integer
      - description: Limit
        in: query
        name: limit
        schema:
          type: integer
      - description: Provider
        in: query
        name: provider
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WateringPlanList'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get all watering plans
      tags:
      - Watering Plan
    post:
      description: Create watering plan
      operationId: create-watering-plan
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/WateringPlanCreate'
        description: Watering Plan Create Request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WateringPlan'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Create watering plan
      tags:
      - Watering Plan
      x-codegen-request-body-name: body
  /v1/watering-plan/{id}:
    delete:
      description: Delete watering plan
      operationId: delete-watering-plan
      parameters:
      - description: Watering Plan ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          content: {}
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Delete watering plan
      tags:
      - Watering Plan
    get:
      description: Get watering plan by ID
      operationId: get-watering-plan-by-id
      parameters:
      - description: Watering Plan ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WateringPlan'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get watering plan by ID
      tags:
      - Watering Plan
    put:
      description: Update watering plan
      operationId: update-watering-plan
      parameters:
      - description: Watering Plan ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/WateringPlanUpdate'
        description: Watering Plan Update Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WateringPlan'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Update watering plan
      tags:
      - Watering Plan
      x-codegen-request-body-name: body
  /v1/watering-plan/route/gpx/{gpx_name}:
    get:
      description: Generate route
      parameters:
      - description: gpx file name
        in: path
        name: gpx_name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/gpx+xml:
              schema:
                format: binary
                type: string
            application/json:
              schema:
                format: binary
                type: string
          description: OK
        "400":
          content:
            application/gpx+xml:
              schema:
                $ref: '#/components/schemas/HTTPError'
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "500":
          content:
            application/gpx+xml:
              schema:
                $ref: '#/components/schemas/HTTPError'
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Generate route
      tags:
      - Watering Plan
  /v1/watering-plan/route/preview:
    post:
      description: Generate preview route
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
        description: Route Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoJson'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Generate preview route
      tags:
      - Watering Plan
      x-codegen-request-body-name: body
components:
  schemas:
...
    EvaluationValue:
      example:
        consumed_water: 1.4658129805029452
        tree_cluster_id: 5
        watering_plan_id: 5
      properties:
        consumed_water:
          type: number
        tree_cluster_id:
          type: integer
        watering_plan_id:
          type: integer
      required:
      - consumed_water
      - tree_cluster_id
      - watering_plan_id
      type: object
...
    GeoJson:
      example:
        features:
        - bbox:
          - 6.027456183070403
          - 6.027456183070403
          geometry:
            coordinates:
            - - 1.4658129805029452
              - 1.4658129805029452
            - - 1.4658129805029452
              - 1.4658129805029452
            type: FeatureCollection
          type: null
          properties:
            key: "{}"
        - bbox:
          - 6.027456183070403
          - 6.027456183070403
          geometry:
            coordinates:
            - - 1.4658129805029452
              - 1.4658129805029452
            - - 1.4658129805029452
              - 1.4658129805029452
            type: FeatureCollection
          type: null
          properties:
            key: "{}"
        metadata:
          end_point:
            latitude: 5.962133916683182
            longitude: 5.637376656633329
          start_point:
            latitude: 5.962133916683182
            longitude: 5.637376656633329
          watering_point:
            latitude: 5.962133916683182
            longitude: 5.637376656633329
        bbox:
        - 0.8008281904610115
        - 0.8008281904610115
        type: null
      properties:
        bbox:
          items:
            type: number
          type: array
        features:
          items:
            $ref: '#/components/schemas/GeoJsonFeature'
          type: array
        metadata:
          $ref: '#/components/schemas/GeoJSONMetadata'
        type:
          $ref: '#/components/schemas/GeoJsonType'
      required:
      - bbox
      - features
      - metadata
      - type
      type: object
...
    GeoJsonType:
      enum:
      - FeatureCollection
      type: string
      x-enum-varnames:
      - FeatureCollection
...
    HTTPError:
      example:
        path: path
        code: 7
        method: method
        error: error
      properties:
        code:
          type: integer
        error:
          type: string
        method:
          type: string
        path:
          type: string
      required:
      - code
      - error
      - method
      - path
      type: object
...
    Pagination:
      example:
        next_page: 9
        total_records: 4
        total_pages: 2
        prev_page: 3
        current_page: 7
      properties:
        current_page:
          type: integer
        next_page:
          type: integer
        prev_page:
          type: integer
        total_pages:
          type: integer
        total_records:
          type: integer
      required:
      - current_page
      - next_page
      - prev_page
      - total_pages
      - total_records
      type: object
...
    RouteRequest:
      example:
        trailer_id: 6
        cluster_ids:
        - 0
        - 0
        transporter_id: 1
      properties:
        cluster_ids:
          items:
            type: integer
          type: array
        trailer_id:
          type: integer
        transporter_id:
          type: integer
      required:
      - cluster_ids
      - transporter_id
      type: object
...
    WateringPlan:
      example:
        date: date
        treeclusters:
        - address: address
          latitude: 6.027456183070403
          created_at: created_at
          description: description
          last_watered: last_watered
          soil_condition: schluffig
          archived: true
          tree_ids:
          - 2
          - 2
          updated_at: updated_at
          watering_status: good
          name: name
          moisture_level: 5.962133916683182
          id: 0
          region:
            name: name
            id: 5
          longitude: 1.4658129805029452
        - address: address
          latitude: 6.027456183070403
          created_at: created_at
          description: description
          last_watered: last_watered
          soil_condition: schluffig
          archived: true
          tree_ids:
          - 2
          - 2
          updated_at: updated_at
          watering_status: good
          name: name
          moisture_level: 5.962133916683182
          id: 0
          region:
            name: name
            id: 5
          longitude: 1.4658129805029452
        cancellation_note: cancellation_note
        distance: 0.8008281904610115
        user_ids:
        - user_ids
        - user_ids
        total_water_required: 9.301444243932576
        refill_count: 7
        created_at: created_at
        description: description
        duration: 6.027456183070403
        evaluation:
        - consumed_water: 1.4658129805029452
          tree_cluster_id: 5
          watering_plan_id: 5
        - consumed_water: 1.4658129805029452
          tree_cluster_id: 5
          watering_plan_id: 5
        trailer:
          water_capacity: 5.962133916683182
          length: 1.4658129805029452
          number_plate: number_plate
          created_at: created_at
          description: description
          weight: 5.637376656633329
          type: transporter
          additional_information:
            key: ""
          driving_license: B
          updated_at: updated_at
          provider: provider
          width: 2.3021358869347655
          model: model
          id: 6
          height: 0.8008281904610115
          status: active
        additional_information:
          key: ""
        updated_at: updated_at
        provider: provider
        gpx_url: gpx_url
        transporter:
          water_capacity: 5.962133916683182
          length: 1.4658129805029452
          number_plate: number_plate
          created_at: created_at
          description: description
          weight: 5.637376656633329
          type: transporter
          additional_information:
            key: ""
          driving_license: B
          updated_at: updated_at
          provider: provider
          width: 2.3021358869347655
          model: model
          id: 6
          height: 0.8008281904610115
          status: active
        id: 2
        status: planned
      properties:
        additional_information:
          additionalProperties: true
          type: object
        cancellation_note:
          type: string
        created_at:
          type: string
        date:
          type: string
        description:
          type: string
        distance:
          type: number
        duration:
          type: number
        evaluation:
          items:
            $ref: '#/components/schemas/EvaluationValue'
          type: array
        gpx_url:
          type: string
        id:
          type: integer
        provider:
          type: string
        refill_count:
          type: integer
        status:
          $ref: '#/components/schemas/WateringPlanStatus'
        total_water_required:
          type: number
        trailer:
          $ref: '#/components/schemas/Vehicle'
        transporter:
          $ref: '#/components/schemas/Vehicle'
        treeclusters:
          items:
            $ref: '#/components/schemas/TreeClusterInList'
          type: array
        updated_at:
          type: string
        user_ids:
          items:
            type: string
          type: array
      required:
      - additional_information
      - cancellation_note
      - created_at
      - date
      - description
      - distance
      - duration
      - evaluation
      - gpx_url
      - id
      - provider
      - refill_count
      - status
      - total_water_required
      - transporter
      - treeclusters
      - updated_at
      - user_ids
      type: object
    WateringPlanCreate:
      example:
        date: date
        additional_information:
          key: ""
        tree_cluster_ids:
        - 1
        - 1
        provider: provider
        user_ids:
        - user_ids
        - user_ids
        trailer_id: 0
        description: description
        transporter_id: 6
      properties:
        additional_information:
          additionalProperties: true
          type: object
        date:
          type: string
        description:
          type: string
        provider:
          type: string
        trailer_id:
          type: integer
        transporter_id:
          type: integer
        tree_cluster_ids:
          items:
            type: integer
          type: array
        user_ids:
          items:
            type: string
          type: array
      required:
      - date
      - description
      - transporter_id
      - tree_cluster_ids
      - user_ids
      type: object
...
    WateringPlanList:
      example:
        pagination:
          next_page: 9
          total_records: 4
          total_pages: 2
          prev_page: 3
          current_page: 7
        data:
        - date: date
          treeclusters:
          - address: address
            latitude: 6.027456183070403
            created_at: created_at
            description: description
            last_watered: last_watered
            soil_condition: schluffig
            archived: true
            tree_ids:
            - 2
            - 2
            updated_at: updated_at
            watering_status: good
            name: name
            moisture_level: 5.962133916683182
            id: 0
            region:
              name: name
              id: 5
            longitude: 1.4658129805029452
          - address: address
            latitude: 6.027456183070403
            created_at: created_at
            description: description
            last_watered: last_watered
            soil_condition: schluffig
            archived: true
            tree_ids:
            - 2
            - 2
            updated_at: updated_at
            watering_status: good
            name: name
            moisture_level: 5.962133916683182
            id: 0
            region:
              name: name
              id: 5
            longitude: 1.4658129805029452
          cancellation_note: cancellation_note
          distance: 0.8008281904610115
          user_ids:
          - user_ids
          - user_ids
          total_water_required: 1.4658129805029452
          created_at: created_at
          description: description
          trailer:
            water_capacity: 5.962133916683182
            length: 1.4658129805029452
            number_plate: number_plate
            created_at: created_at
            description: description
            weight: 5.637376656633329
            type: transporter
            additional_information:
              key: ""
            driving_license: B
            updated_at: updated_at
            provider: provider
            width: 2.3021358869347655
            model: model
            id: 6
            height: 0.8008281904610115
            status: active
          updated_at: updated_at
          transporter:
            water_capacity: 5.962133916683182
            length: 1.4658129805029452
            number_plate: number_plate
            created_at: created_at
            description: description
            weight: 5.637376656633329
            type: transporter
            additional_information:
              key: ""
            driving_license: B
            updated_at: updated_at
            provider: provider
            width: 2.3021358869347655
            model: model
            id: 6
            height: 0.8008281904610115
            status: active
          id: 6
          status: planned
        - date: date
          treeclusters:
          - address: address
            latitude: 6.027456183070403
            created_at: created_at
            description: description
            last_watered: last_watered
            soil_condition: schluffig
            archived: true
            tree_ids:
            - 2
            - 2
            updated_at: updated_at
            watering_status: good
            name: name
            moisture_level: 5.962133916683182
            id: 0
            region:
              name: name
              id: 5
            longitude: 1.4658129805029452
          - address: address
            latitude: 6.027456183070403
            created_at: created_at
            description: description
            last_watered: last_watered
            soil_condition: schluffig
            archived: true
            tree_ids:
            - 2
            - 2
            updated_at: updated_at
            watering_status: good
            name: name
            moisture_level: 5.962133916683182
            id: 0
            region:
              name: name
              id: 5
            longitude: 1.4658129805029452
          cancellation_note: cancellation_note
          distance: 0.8008281904610115
          user_ids:
          - user_ids
          - user_ids
          total_water_required: 1.4658129805029452
          created_at: created_at
          description: description
          trailer:
            water_capacity: 5.962133916683182
            length: 1.4658129805029452
            number_plate: number_plate
            created_at: created_at
            description: description
            weight: 5.637376656633329
            type: transporter
            additional_information:
              key: ""
            driving_license: B
            updated_at: updated_at
            provider: provider
            width: 2.3021358869347655
            model: model
            id: 6
            height: 0.8008281904610115
            status: active
          updated_at: updated_at
          transporter:
            water_capacity: 5.962133916683182
            length: 1.4658129805029452
            number_plate: number_plate
            created_at: created_at
            description: description
            weight: 5.637376656633329
            type: transporter
            additional_information:
              key: ""
            driving_license: B
            updated_at: updated_at
            provider: provider
            width: 2.3021358869347655
            model: model
            id: 6
            height: 0.8008281904610115
            status: active
          id: 6
          status: planned
      properties:
        data:
          items:
            $ref: '#/components/schemas/WateringPlanInList'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
      - data
      type: object
...
    WateringPlanUpdate:
      example:
        date: date
        evaluation:
        - consumed_water: 1.4658129805029452
          tree_cluster_id: 5
          watering_plan_id: 5
        - consumed_water: 1.4658129805029452
          tree_cluster_id: 5
          watering_plan_id: 5
        additional_information:
          key: ""
        cancellation_note: cancellation_note
        tree_cluster_ids:
        - 1
        - 1
        provider: provider
        user_ids:
        - user_ids
        - user_ids
        trailer_id: 0
        description: description
        transporter_id: 6
        status: planned
      properties:
        additional_information:
          additionalProperties: true
          type: object
        cancellation_note:
          type: string
        date:
          type: string
        description:
          type: string
        evaluation:
          items:
            $ref: '#/components/schemas/EvaluationValue'
          type: array
        provider:
          type: string
        status:
          $ref: '#/components/schemas/WateringPlanStatus'
        trailer_id:
          type: integer
        transporter_id:
          type: integer
        tree_cluster_ids:
          items:
            type: integer
          type: array
        user_ids:
          items:
            type: string
          type: array
      required:
      - cancellation_note
      - date
      - description
      - status
      - transporter_id
      - tree_cluster_ids
      - user_ids
      type: object
...
  securitySchemes:
    Keycloak:
      flows:
        authorizationCode:
          authorizationUrl: https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/auth
          scopes: {}
          tokenUrl: https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/token
      type: oauth2
...